refactor(Machine Learning Plot) Refatorando componente para melhor componentização.
