+++ src/styles/machine-learning-plot.css
@@ -1,0 +1,60 @@
+.machine_learning {
+    display: flex;
+    width: 100%;
+    max-width: 1440px;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.machine_learning_info {
+    margin-top: 45px;
+    margin-left: 50px;
+}
+
+.machine_learning_title {
+    font-size: 2rem;
+    font-weight: 500;
+    margin-bottom: 0.5em;
+    color: var(--secondaryColor);
+}
+
+.machine_learning_description {
+    white-space: pre-line;
+    width: 650px;
+    font-size: 1.1rem;
+    text-align: justify;
+    color: var(--textColor);
+    margin-top: 30px;
+}
+
+@media only screen and (max-width: 980px) {
+    .machine_learning {
+        flex-direction: column;
+        align-items: center;
+    }
+    .machine_learning_info {
+        width: 80%;
+        margin-top: 0px;
+        display: flex;
+        flex-direction: column;
+        justify-content: space-between;
+        align-items: center;
+        margin-bottom: 30px;
+        margin-left: 0px;
+        text-align: center;
+    }
+    .machine_learning_description {
+        width: 100%;
+        height: auto;
+        margin-bottom: 30px;
+    }
+}
+
+@media only screen and (max-width: 600px) {
+    .machine_learning_info_title {
+        font-size: 1.5rem;
+        font-weight: 500;
+        margin-bottom: 0.5em;
+        color: var(--secondaryColor);
+    }
+}
